% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combits.R
\name{combits}
\alias{combits}
\title{combits}
\usage{
combits(num, bit = 5)
}
\arguments{
\item{num:}{A numeric vector.}

\item{bit:}{The bit number.}
}
\value{
A completed character vector.
}
\description{
Complete bits.
}
\examples{
\dontrun{
combits(num = seq(1, 100, 1), bit = 5)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}

